extracted queue and stack to separate template classes
