+++ src/pattern_library/pattern_sort_merge.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../pattern_internals/sorter_pattern.h"
+#include "../data_structures/stack.h"
 
 // first, some helper classes/structs/enums for merge sort
 
@@ -19,11 +20,8 @@ typedef struct RequiredAction_ {
  * A data structure for keeping track of actions performed by Merge Sort
  * (analogous to the call stack in a normal recursive merge sort)
  */
-class PendingStack {
+class PendingStack : public Stack<RequiredAction> {
     public:
-        PendingStack();
-        ~PendingStack();
-
         /**
          * Add a split action to the stack
          * @param start the start index of the split action
@@ -37,28 +35,6 @@ class PendingStack {
          * @param end   the end index of the merge action
          */
         void pushMergeAction(int start, int mid, int end);
-
-        /**
-         * Get the next required action
-         * @return pointer to the action on top, or NULL if stack is empty
-         */
-        RequiredAction* pop();
-
-        /**
-         * Determine whether stack has any elements
-         * @return true if stack has size >= 1, false if stack is empty
-         */
-        bool nonempty();
-    private:
-        typedef struct ListNode_ {
-            RequiredAction* element;
-            struct ListNode_* next;
-        } ListNode;
-
-        void push(RequiredAction* action);
-
-        ListNode* top;
-        unsigned int size;
 };
 
 class Pattern_MergeSort: public SorterPattern {
