+++ src/pattern_library/pattern_sort_quick.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../pattern_internals/sorter_pattern.h"
+#include "../data_structures/queue.h"
 
 // first, some helper classes/structs/enums for merge sort
 
@@ -13,41 +14,14 @@ typedef struct RequiredSort_ {
 /**
  * A queue for keeping track of actions performed by Quick Sort
  */
-class PendingQueue {
+class PendingQueue : public Queue<RequiredSort> {
     public:
-        PendingQueue();
-        ~PendingQueue();
-
         /**
          * Add a required sort to the queue
          * @param start the start index of the sort action
          * @param end   the end index of the sort action
          */
         void pushPendingSort(int start, int end);
-
-        /**
-         * Get the next required action
-         * @return pointer to the action on top, or NULL if queue is empty
-         */
-        RequiredSort* pop();
-
-        /**
-         * Determine whether queue has any elements
-         * @return true if queue has size >= 1, false if queue is empty
-         */
-        bool nonempty();
-    private:
-        typedef struct ListNode_ {
-            RequiredSort* element;
-            struct ListNode_* next;
-        } ListNode;
-
-        void push(RequiredSort* action);
-
-        ListNode* front;
-        ListNode* back;
-
-        unsigned int size;
 };
 
 class Pattern_QuickSort: public SorterPattern {
